+++ src/components/certificates/types.ts
@@ -1,0 +1,10 @@
+export type CertificatesFetchingStatus = "pending" | "resolved" | "rejected";
+
+export interface CertificatesFetchingType {
+  connectionClientUpdate?: CertificatesFetchingStatus;
+  connectionDetect?: CertificatesFetchingStatus;
+  connectionSupport?: CertificatesFetchingStatus;
+  connectionApprove?: CertificatesFetchingStatus;
+  providers?: CertificatesFetchingStatus;
+  certificates?: CertificatesFetchingStatus;
+}
